# Add SHA512/224 and SHA512/256 benchmarks