--- a/wolfcrypt/benchmark/benchmark.h
+++ b/wolfcrypt/benchmark/benchmark.h
@@ -67,6 +67,8 @@ void bench_sha224(int useDeviceID);
 void bench_sha256(int useDeviceID);
 void bench_sha384(int useDeviceID);
 void bench_sha512(int useDeviceID);
+void bench_sha512_224(int useDeviceID);
+void bench_sha512_256(int useDeviceID);
 void bench_sha3_224(int useDeviceID);
 void bench_sha3_256(int useDeviceID);
 void bench_sha3_384(int useDeviceID);